extracted an abstract base class
