+++ src/main/java/org/cicirello/search/evo/FitnessProportionateSelection.java
@@ -37,22 +37,12 @@ import java.util.concurrent.ThreadLocalRandom;
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
-public class FitnessProportionateSelection implements SelectionOperator {
+public class FitnessProportionateSelection extends AbstractFitnessWeightedSelection {
 	
 	/**
 	 * Construct a fitness proportionate selection operator.
 	 */
 	public FitnessProportionateSelection() {}
-	
-	@Override
-	public final void select(PopulationFitnessVector.Integer fitnesses, int[] selected) {
-		selectAll(normalizeWeights(computeWeightRunningSum(fitnesses)), selected);
-	}
-	
-	@Override
-	public final void select(PopulationFitnessVector.Double fitnesses, int[] selected) {
-		selectAll(normalizeWeights(computeWeightRunningSum(fitnesses)), selected);
-	}
 	
 	@Override
 	public FitnessProportionateSelection split() {
@@ -64,54 +54,22 @@ public class FitnessProportionateSelection implements SelectionOperator {
 	/*
 	 * package private to enable subclasses in same package to override
 	 */
-	double[] computeWeightRunningSum(PopulationFitnessVector.Integer fitnesses) {
-		double[] p = new double[fitnesses.size()];
-		p[0] = fitnesses.getFitness(0);
-		for (int i = 1; i < p.length; i++) {
-			p[i] = p[i-1] + fitnesses.getFitness(i);
-		}
-		return p;
-	}
-	
-	/*
-	 * package private to enable subclasses in same package to override
-	 */
-	double[] computeWeightRunningSum(PopulationFitnessVector.Double fitnesses) {
-		double[] p = new double[fitnesses.size()];
-		p[0] = fitnesses.getFitness(0);
-		for (int i = 1; i < p.length; i++) {
-			p[i] = p[i-1] + fitnesses.getFitness(i);
-		}
-		return p;
-	}
-	
-	/*
-	 * package private to enable subclasses in same package to override
-	 */
-	void selectAll(double[] weights, int[] selected) {
+	@Override
+	void selectAll(double[] normalizedWeights, int[] selected) {
 		for (int i = 0; i < selected.length; i++) {
-			selected[i] = selectOne(weights, 0, weights.length-1, ThreadLocalRandom.current().nextDouble());
+			selected[i] = selectOne(normalizedWeights, 0, normalizedWeights.length-1, ThreadLocalRandom.current().nextDouble());
 		}
 	}
 	
-	private int selectOne(double[] weights, int first, int last, double u) {
+	private int selectOne(double[] normalizedWeights, int first, int last, double u) {
 		if (last <= first) {
 			return first;
 		}
 		int mid = (first + last) >> 1;
-		if (u < weights[mid]) {
-			return selectOne(weights, first, mid, u);
+		if (u < normalizedWeights[mid]) {
+			return selectOne(normalizedWeights, first, mid, u);
 		} else {
-			return selectOne(weights, mid+1, last, u);
+			return selectOne(normalizedWeights, mid+1, last, u);
 		}
-	}
-	
-	private double[] normalizeWeights(double[] weights) {
-		double total = weights[weights.length-1];
-		weights[weights.length-1] = 1.0;
-		for (int i = weights.length-2; i >= 0; i--) {
-			weights[i] /= total;
-		}
-		return weights;
 	}
 }
